Separate internal file layer into separate package
