+++ internal/vfs/osfs/common_unix.go
@@ -1,0 +1,5 @@
+// +build darwin dragonfly freebsd linux netbsd openbsd solaris
+
+package osfs
+
+type osFileState struct{}
